Conflicts: 	app/mapping.txt

+++ app/src/main/java/com/dkhs/portfolio/ui/MyPurseActivity.java
@@ -47,7 +47,7 @@ public class MyPurseActivity extends ModelAcitivity implements View.OnClickListe
     private TextView mBalanceOutTv;
 
     private boolean withDrawAvailable = false;
-    private float available = 0;
+    private double available = 0;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
